Add distances to nearby zctas, revert previous commit
I was going to replace in the call to ST_BUFFER within mysql query and instead
use RGeo's #buffer method, but it misses some points. We'll just have to live
with slightly less readable code.

I wasn't sure whether the distance calculations should go in the model or in the
controller, but it seems more coherent in the model for now.