Temporary fix of scaling issues
Turns out my units of measure are not meters, but rather the map units of Web
Mercator. This is off by a factor of cos(latitude). For now, I am fixing this
problem by adjusting by the scale factor. This is fine for small distances but
is bad for long distances where the latitude changes significantly. It's also
bad since straight lines on the Mercator projection are NOT straight lines on
the globe (and vice versa).